feat: The program now stores the clients data from the clients.csv file
Here begins the development of the bank account management system
program. Many of the source files currently in the folder of this
project were included from the "Blockbuster" project, since they might
be useful during development. These source files are likely to be
heavily modified or otherwise deleted during development.